requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
qwen-vl-utils>=1.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
wandb>=0.15.0
hydra-core>=1.3.0
omegaconf>=2.3.0
einops>=0.7.0
timm>=0.9.0
albumentations>=1.3.0
monai>=1.3.0
nibabel>=5.1.0
pydicom>=2.4.0
SimpleITK>=2.3.0
pillow>=10.0.0
scipy>=1.11.0
scikit-image>=0.21.0
tensorboard>=2.14.0
rich>=13.5.0
PyYAML>=6.0